backend/requirements.txt
# RecapVideo.AI Backend
# Python 3.11+

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache & Queue (Redis - for Celery workers)
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.13.3

# TTS & Processing
edge-tts>=7.2.7
gTTS>=2.5.0
yt-dlp>=2025.12.8
pytubefix>=6.0.0
ffmpeg-python==0.2.0
webvtt-py==0.4.6
curl_cffi>=0.13.0,<0.14.0

# External APIs
google-generativeai>=0.8.0
groq>=0.11.0
resend==0.7.2
boto3==1.34.34

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3
user-agents==2.2.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.3.0
ruff==0.1.14
